(Don't stage) Starting from working code
This commit is just for showing my process, I'm not going to stage this as it doesn't make real improvements. Much of the new CSS and JS added in this commit will be pulled out again or edited in the next commit. This is because I'm just adding in some working code and useful style rules from elsewhere, I'll pull out anything we don't need and alter it to match the project, but I wanted a commit with raw stuff I copied in, so that the comparison between commits will show what I've changed.