Started actually building game + renamed mainStyles
During this I realised I needed two separate major CSS files - one for the actual game, and one for the general purpose stuff that is in the landing page as well. (For "Separation of Concerns"). So I created gameStyles.css and renamed mainStyles.css to generalStyles.css (in the hope that the purpose is more clear)

Game isn't playable at this point yet, but I have put in a game area, an element that can stand in for a character while I figure out movement, and buttons for movement that all call a function move(dir) - passing in a letter to represent the direction of movement