Working game boundary!
Oh man that was some effort coding this!

This update was all about getting game boundary collision detection working.

Lots of small changes that make up this update - in the order that github shows the files:
a) To make it visually clear while testing, I changed our stand in character we're moving. I made the div have a set width and height and gave it a red bg colour, and changed the markup to be an empty div instead of the word "yo"

b)I added a section of globals (our global JSON object holding game values) for our player character, since width and height of it where needed - actually those two values probably don't need to be global vars thinking about it now, but we will need character stats as global things

c)In last commit I tested the new isInBounds function with a line at the top of move - that isn't needed in test form now as it's used for real checks further down
d)Noticed that offsetTop is off by 2 - previously it wasn't obvious, but now that we set both x and y co-ords it would be obvious - in testing left and right movement actually went diagonal
e)renamed attemptX and attemptY because further down the code we need values for each edge for collision detection
f)now  we don't physically make the moves in the section with ifs for each direction, we just change the attempt value. Then once that's done, we work out right side and bottom axis values, put the axis values for each edge into isInBounds and if true use attempt values to reposition the PC
g)Seems like when I made the isInBounds function I confused the x and y axes around so had to rename variables