Beginnings of keyboard controls + better info sizes
The ramble in the alpha area was just way too long, nobody would bother reading it, and it was kinda in the way and taking too much of my time keeping that up to date. So I reduced alot of it down, moved some of the message I was trying to convey onto the landing page - altering it for the destination and made the key text that gets updated in the alpha area orange and in a consise format.