all buttons now continuously move on hold
Timing might be different between keys and buttons. I've tried to make it the same speed, but keys seem like they might accelerate rather than have constant motion, or maybe a sudden change in speed after a second. For now this will do, but if it effects playtesting or I think of an easy solution I can change it later.