Beta 0.0.3 "Awaken in a spin"
The highlight of this update is that throughout the opening sequence, the rainbow vortex now spins on it's own. (This is quite a big improvement, it's a much more trippy and mysterious start)

Also added a note on screen002 to click text to move it to the next screen.
Next update I think I'll add an automatic move on from screen001