Beta 0.0.5 "Fix for rainbow lag" (also fixed screen002 jump back)
Rainbow vortex being a background image applied in CSS, seems to have caused a visible lag when run online rather than from local file. Note it took quite a bit of trial and improvement tweaking the image and CSS to get it to look roughly the same as it did.

It's now using a file that is just the central part - since a large amount of it was clipped off. I've kept the old file (though with some changes I made to help identify corners on the rotating image), in case it's useful at some point.

Also fixes a bug that was introduced recently - the automatic jump to screen002 didn't check you where still on screen001 - meaning if you where fast to get to screen003 or beyond, screen002 would be shown as well as whatever screen you where on. The code to fix this also means we are now  actually storing the screen being transitioned to during the nextScreen function in the var globals.Screens.current