Acknowledging the styling issues on mobile
At this point I don't know if I'll be able to fix up all the CSS issues for mobile today or the next couple of days, but I've not really stated those issues anywhere. So this covers me for a bit, and lets anyone visiting know that I'll be working on it and getting that fixed soon, without locking in a date that might be hard to keep.