wonky boundaries likely due to cache or file not updated on server
The last commit had each approach to preventing the rainbow going out of the playarea - the one that had issues in prod was in the test area, and the always working kludge in beta area - both worked then. So presumably the issue was down to a cached file or maybe the server hadn't fully update the file.

This commit alters the beta area back to using the  non-kludgey version and also brings in use of query strings on files - supposedly that's a good front end fix for cacheing