Testing: Scrollable start screen, less story on screen at once, smaller headings except on big screens
Partially fixed an issue on the start screen ("screen002") where you couldn't scroll down to keep reading content that didn't fit on the screen. (the issue is still there in the scenario that a user goes past that screen and uses the previous button to go back to it - that is related to changing between full size screen)
Opening screens are now in a new class called openingScreen rather than in storyTxt - so that the opening screens aren't fixed position.
Removed the onclick for "screen002" so that people on touch screens can flick the screen up and down without triggering it - there's a next button by that point anyway.
Removed the "to continue from this screen ..... tap the text" message on the start screen - it's not needed anymore and would be misleading now that I've removed the onclick event from that screen.

Less story on screen at once - to help reduce text overlapping game area

Smaller headings except on big screens - Helps fit more text above the fold on smaller screens