Testing: Text at top during story given big space on taller screens
This helps a few things
- further mitigating possibility of browser full screen message overlapping start of story
 - Keeps the background behind the text stay in a constant spot, which looks much nicer
- Lets the text be a little nearer to centre on larger screens - also quite a bit nicer

The solution implemented here doesn't apply the changes to the text on screen during gameplay - those are now on another class that still gets the initial CSS but not the increased sizing MediaQuery since we want that text to be tucked out of the way of gameplay - maybe at an even bigger screen size we could apply similar MediaQuery rules to the new class. For the sizes I'm perfecting at the moment this works well.

Also moved a line of story back one screen to a previously empty screen - for better pacing as you're reading it.