Test area: keeping the bigger game area
The cause of the wall divs being moved when you carry on with the next section of the game, was changing the game area size. I made the game area shrink back down after playable sections because that felt right with the initial plan for movement sections being very seperate to story cut scene sections. Now that isn't the case, it doesn't really make sense to shrink the game area anyway. Now we have walls shown, no game area shrink and I'm intending on having the Dr position shown when they enter, it felt necessary to still adjust where the player is when we get to screen011  but slightly off from where it was being moved to. It also felt needed to add a story screen explaining to the player why they've moved