Test area: can you hear choice neatening up
I'm gearing towards releasing to Beta, with this update ending on the first choices screen with some overlay saying that it's finished, that lets them see the choices but not click them, so it's like a tease.

So this update, I tidied up that choices screen a little. It looked okay on laptop sized screens already but had some overlap and overflow issues on mobile sized screens, plus even on laptop size screens 5 choices seems a bit too much for right now, particularly as each choice leads to some kind of branching and I don't want a huge gap between people playing this beta version and being able to choose any of the options that they can see.

Detail:
 - Temporarily commented out the "Yes, though I can't see you" option, I also commented an idea for how to neatly have more options in future.
 - put a line break where mobile tends to split up the options anyway - I think that helps distinguish it
- added a line break within the crazy option and also gave it a new class for double line buttons (note they didn't need to be double the height, just a bit bigger - there's less space around the text, but it looks right) - line break picks a better spot to split than it was splitting for mobile and also saves it from being really wide

Also added some more improvement ideas that I've had recently