Test area: updated control info
Though this commit only effects a few lines of html, it took quite a bit of thought and experimentation to realise this was the best option for making the keypad controls consistent with the main game. I looked into implementing keyHints throughout the game, but they actually seemed to add complexity for the player and developer, wouldn't have been usable on screens with other typing, and would have been very time consuming and messy to fully implement.