WIP: planning out Asylum Builder Tools,
I'm going to create a tool for making rooms and putting them into the grid. This will start out as something for me, but once it's refined a bit I'll include links from the game home page and make it a key feature. A HUGE part of SRA is the concept of encouraging people to get creative and get comfortable contributing to a community effort thing.