(Just for reference) layout tool default JSON used for eg
How would our actual JS code look if we replaced what we had made by hand at this point with what the layout tool creates by this early stage?

This commit is the answer (that I will put back straight after). It's not as unreadable as it could be, but still not all that readable, and it's already quite a lot of code to read through. 100 quite long lines that are each difficult to read. I'm unsure if it's fully valid JS , the game seems to work fine up to the point where you attempt escape, but I can't remember the state of the game before pasting this all in, so don't know (until I revert) if this is the cause.