Update comments in generalStyles.css
I spotted that there where some comments on style rules, saying they would be "potentially useful" that I know to be in use - felt that was misleading now, so I went through checking all of the styles in that file.
Also reminded me of some classes that could be useful.