WIP Asylum Tool: Able to set designers notes
NB not safe for actually running on a live webserver right now!!! XSS attacks likely to be possible.  This is a great start for making the asylum editable.

Whilst doing this I discovered a major bug that only just now took effect but was in the default asylum creation JS. I noticed that editing one thing edited it everywhere - the old code for copying the default room 100 times actually just put in a link to that object - so any edits on one edited the default room, which all of the others referenced.