wip Asylum Tool:
A simple change for this update, but a really great improvement for my experience as a user of this tool. I realised that one of the great things about colour coding by label was that the colour coding gave very quick progress feedback. That felt missing with designers notes - I think due to more clicks. I made the design completeness select trigger the function for applying changes with an onchange event