WIP Asylum tool: nextCell data
nextCell data will tell the game where to go when a player character goes through a bound. This will replace doorTo - we need this info for every bound - in case the character breaks a wall or some other unexpected way of getting through where there isn't a door, and this implementation gives us an automatically generated default for every bound at Asylum design time, that we'll be able to override at Asylum design time!