wip Asylum Tool: save editor bounds data
Boo yeah! now it saves the main bounds data!

There's still some data that isn't being saved but we can now adjust the bounds types of cells in the editor, and the values - although the values are a bit awkward to adjust since there's no quick visual feedback - there's no way to see how it will look in game from within the editor yet, so you'd have to load the JSON into the game and test it