delving into building the asylum Layout
I'm at a point now with the Asylum Tool where it's made it quite easy and vaguely compelling for a minute or two to go through and set what kind of walls are in each room. So I've done a chunk of that for this little commit.  The real gold here is actually the visual feedback you get from the grid. Seeing the grid gradually look more and more like a map of an Asylum really feels like I'm making progress. It's a bit like the compelling feeling of a jigsaw. There are a couple of improvements that could help speed it up (and therefore make the feeling of progress better, for what I think will be a virtuous cycle)