WIP Asylum tool: yet more work on node choosing
This is quite a tricky bit of code to do - or at least its challenging me. So I spent some time thinking through in simple terms how it would have to work - or at least one implementation that came to mind. I got the first part of that done here