Asylum tool: one step node choosing
Since this is proving more difficult than I had anticipated, and I'd like to have this looking tidy and be able to shift to other areas of the project, I'm changing my approach to building the node chooser to a better iterative approach. Rather than building all of the JSON drilling down part, and then building all of the JSON showing/editing part, this update has a nice clean single select that lets you choose the top JSON node you want, and that choice is displayed below. The JSON isn't editable yet, but we'll add some ability to do that in another update.

Also for now, we're not calling the function to create more selects or number input anymore. We'll add in the next layer down in another update. It may even make sense for the next update for drilling down, to just add in drilling down into rooms, since that will be really helpful whereas drilling into the other nodes isn't as much.